abstract out createDropdownControl
